Add files via upload
A correlation heatmap between our variables. Our heatmap indicates that none of the feature attributes are strongly correlated with our target variable, 'dpnm.'
We also notice multicollinearity between the PAY and BILL attributes, as we would expect intuitively. However, because this is a classification model (as opposed to a regression model), this is not as critical because our primary goal is prediction and not interpretation.